// apps/web/src/lib/api/types.ts
export type DryRunDescriptor = {
  action: string;
  wallet_name: string;
  wallet_pubkey: string;
  intent_index: number;
  intent_pubkey: string;
  message_hex: string;
  params_data_hex: string;
  expiry: number;
  proposal_pubkey?: string;
  proposal_index?: number;
};